Created custom Error Messages for different Error Types using ErrorOr
This description is not related to this commit but please check the backend.csproj file, the author forgot to separate commit "Add Contract Service". In this earlier commit, a project reference to the Contracts project is created.

The ErrorOr Package was added as it provides a fluent API for handling errors and results in .NET 6 Web API applications.